+++ src/language/server.ts
@@ -9,6 +9,7 @@ import { Client } from 'pg';
 import * as fs from 'fs';
 import { Validator } from './validator';
 import { IConnection as IDBConnection } from '../common/IConnection';
+import { BackwardIterator } from '../common/backwordIterator';
 
 export interface ISetConnection { 
   connection: IDBConnection
@@ -59,6 +60,7 @@ export interface DBFunction {
 
 let tableCache: DBTable[] = [];
 let functionCache: DBFunction[] = [];
+let keywordCache: string[] = [];
 
 /**
  * To Debug the language server
@@ -89,16 +91,19 @@ connection.onInitialize((_params) : InitializeResult => {
         triggerCharacters: [' ', '.']
       },
       signatureHelpProvider: {
-        triggerCharacters: ['(']
+        triggerCharacters: ['(', ',']
       }
     }
   }
 });
 
-function dbConnectionEnded() {
-  dbConnection = null;
-}
-
+/*
+.dP"Y8 888888 888888 88   88 88""Yb     8888b.  88""Yb      dP""b8    db     dP""b8 88  88 888888 
+`Ybo." 88__     88   88   88 88__dP      8I  Yb 88__dP     dP   `"   dPYb   dP   `" 88  88 88__   
+o.`Y8b 88""     88   Y8   8P 88"""       8I  dY 88""Yb     Yb       dP__Yb  Yb      888888 88""   
+8bodP' 888888   88   `YbodP' 88         8888Y"  88oodP      YboodP dP""""Yb  YboodP 88  88 888888 
+*/
+function dbConnectionEnded() { dbConnection = null; }
 async function setupDBConnection(connectionOptions: IDBConnection, uri: string): Promise<void> {
   if (connectionOptions) {
     // dbConnection = await Database.createConnection(conn);
@@ -131,7 +136,6 @@ async function setupDBConnection(connectionOptions: IDBConnection, uri: string):
     
     try {
       if (connectionOptions.database) {
-        console.log('Loading Table/Column Cache');
         let tablesAndColumns = await dbConnection.query(`
           SELECT
             tablename,
@@ -159,7 +163,6 @@ async function setupDBConnection(connectionOptions: IDBConnection, uri: string):
     }
   
     try {
-      console.log('Loading Function Cache');
       let functions = await dbConnection.query(`
         SELECT n.nspname as "schema",
           p.proname as "name",
@@ -202,6 +205,14 @@ async function setupDBConnection(connectionOptions: IDBConnection, uri: string):
       console.log(err);
     }
 
+    try {
+      let keywords = await dbConnection.query(`select * from pg_get_keywords();`);
+      keywordCache = keywords.rows.map<string>(rw => rw.word.toLocaleUpperCase());
+    }
+    catch (err) {
+      console.log(err);
+    }
+
     if (uri) {
       let document = documents.get(uri);
       if (document && document.languageId === 'postgres') {
@@ -220,10 +231,16 @@ connection.onRequest('set_connection', async function() {
   }
   setupDBConnection(newConnection.connection, newConnection.documentUri)
   .catch(err => {
-    console.log('GOT ERROR HERE: ' + err.message)
+    console.log(err.message)
   });
 });
 
+/*
+ dP"Yb  88   88 888888 88""Yb Yb  dP     Yb    dP    db    88     88 8888b.     db    888888 88  dP"Yb  88b 88 
+dP   Yb 88   88 88__   88__dP  YbdP       Yb  dP    dPYb   88     88  8I  Yb   dPYb     88   88 dP   Yb 88Yb88 
+Yb b dP Y8   8P 88""   88"Yb    8P         YbdP    dP__Yb  88  .o 88  8I  dY  dP__Yb    88   88 Yb   dP 88 Y88 
+ `"YoYo `YbodP' 888888 88  Yb  dP           YP    dP""""Yb 88ood8 88 8888Y"  dP""""Yb   88   88  YbodP  88  Y8 
+*/
 documents.onDidOpen((change) => {
   validateTextDocument(change.document);
 });
@@ -232,6 +249,7 @@ documents.onDidChangeContent((change) => {
   validateTextDocument(change.document);
 });
 
+var _NL = '\n'.charCodeAt(0);
 async function validateTextDocument(textDocument: TextDocument): Promise<void> {
   let diagnostics: Diagnostic[] = [];
   // parse and find issues
@@ -243,6 +261,21 @@ async function validateTextDocument(textDocument: TextDocument): Promise<void> {
     }
     for (let sql of Validator.prepare_sql(sqlText)) {
       if (!sql.statement) continue;
+      let errColumnMod = 0;
+      if (sql.statement.trim().toUpperCase().startsWith('EXPLAIN ')) {
+        let match = sql.statement.match(/\s*?EXPLAIN\s/gi);
+        if (match) {
+          for (let i = 0; i < match[0].length; i++) {
+            let ch = match[0].charCodeAt(i);
+            errColumnMod++;
+            if (ch === _NL) {
+              errColumnMod = 1;
+              sql.line++;
+            }
+          }
+          sql.statement = sql.statement.replace(/\s*?EXPLAIN\s/gi, '');
+        }
+      }
       try {
         const results = await dbConnection.query(`EXPLAIN ${sql.statement}`);
       }
@@ -250,7 +283,7 @@ async function validateTextDocument(textDocument: TextDocument): Promise<void> {
         // can use err.position (string)
         // corresponds to full position in query "EXPLAIN ${sql.statement}"
         // need to parse out where in parsed statement and lines that it is
-        let errPosition = parseInt(err.position) - 9; // removes "EXPLAIN " and turn to zero based
+        let errPosition = parseInt(err.position) - 9 + errColumnMod; // removes "EXPLAIN " and turn to zero based
         let errLine = 0;
         while (errPosition > sql.lines[errLine].length) {
           errPosition -= (sql.lines[errLine].length + 1);
@@ -283,21 +316,16 @@ async function validateTextDocument(textDocument: TextDocument): Promise<void> {
   connection.sendDiagnostics({uri: textDocument.uri, diagnostics});
 }
 
+/*
+ dP""b8  dP"Yb  8888b.  888888      dP""b8  dP"Yb  8b    d8 88""Yb 88     888888 888888 88  dP"Yb  88b 88 
+dP   `" dP   Yb  8I  Yb 88__       dP   `" dP   Yb 88b  d88 88__dP 88     88__     88   88 dP   Yb 88Yb88 
+Yb      Yb   dP  8I  dY 88""       Yb      Yb   dP 88YbdP88 88"""  88  .o 88""     88   88 Yb   dP 88 Y88 
+ YboodP  YbodP  8888Y"  888888      YboodP  YbodP  88 YY 88 88     88ood8 888888   88   88  YbodP  88  Y8 
+*/
 connection.onCompletion((e: any): CompletionItem[] => {
-
-  /*
-  p.context.triggerKind:number = 1?
-  p.context.triggerKind 2 - via character
-    triggerCharacter = '.' or whatever
-  p.position
-    .line: number
-    .character: number
-
-  p.textDocument.uri: string
-  */
   let items: CompletionItem[] = [];
   let tableFound = false;
-  if (e.context.triggerKind === 2 && e.context.triggerCharacter === '.') {
+  if (e.context.triggerCharacter === '.') {
     // triggered via . Character
     // look back and grab the text immediately prior to match to table
     let document = documents.get(e.textDocument.uri);
@@ -329,54 +357,71 @@ connection.onCompletion((e: any): CompletionItem[] => {
         kind: CompletionItemKind.Class
       });
       table.columns.forEach(field => {
-        items.push({
-          label: field.attname,
-          kind: CompletionItemKind.Property,
-          detail: field.data_type,
-          documentation: table.tablename
-        });
+        let foundItem = items.find(i => i.label === field.attname && i.kind === CompletionItemKind.Property && i.detail === field.data_type);
+        if (foundItem) {
+          foundItem.documentation += `, ${table.tablename}`;
+        } else {
+          items.push({
+            label: field.attname,
+            kind: CompletionItemKind.Property,
+            detail: field.data_type,
+            documentation: table.tablename
+          });
+        }
       });
     });
     functionCache.forEach(fn => {
       items.push({
         label: fn.name,
         kind: CompletionItemKind.Function,
-        detail: fn.result_type
+        detail: fn.result_type,
+        documentation: fn.overloads[0].description
+      });
+    });
+    keywordCache.forEach(keyword => {
+      items.push({
+        label: keyword,
+        kind: CompletionItemKind.Keyword
       });
     });
   }
   return items;
 });
 
-// connection.onCompletionResolve((item: CompletionItem): CompletionItem => {
-//   item.detail = `${item.label} details`;
-//   item.documentation = `${item.label} documentation`;
-//   return item;
-// });  SignatureHelp
-
+/*
+.dP"Y8 88  dP""b8 88b 88    db    888888 88   88 88""Yb 888888 
+`Ybo." 88 dP   `" 88Yb88   dPYb     88   88   88 88__dP 88__   
+o.`Y8b 88 Yb  "88 88 Y88  dP__Yb    88   Y8   8P 88"Yb  88""   
+8bodP' 88  YboodP 88  Y8 dP""""Yb   88   `YbodP' 88  Yb 888888 
+*/
 connection.onSignatureHelp((positionParams): SignatureHelp => {
   let document = documents.get(positionParams.textDocument.uri);
   let activeSignature = null, activeParameter = null, signatures: SignatureInformation[] = [];
   if (document) {
-    let line = document.getText({
-      start: {line: positionParams.position.line, character: 0},
-      end: {line: positionParams.position.line, character: positionParams.position.character}
+    let iterator = new BackwardIterator(document, positionParams.position.character - 1, positionParams.position.line);
+
+    let paramCount = iterator.readArguments();
+    if (paramCount < 0) return null;
+
+    let ident = iterator.readIdent();
+    if (!ident) return null;
+
+    let fn = functionCache.find(f => f.name.toLocaleLowerCase() === ident.toLocaleLowerCase());
+    if (!fn) return null;
+
+    let overloads = fn.overloads.filter(o => o.args.length >= paramCount);
+    if (!overloads || !overloads.length) return null;
+
+    overloads.forEach(overload => {
+      signatures.push({
+        label: `${fn.name}( ${overload.args.join(' , ')} )`,
+        documentation: overload.description,
+        parameters: overload.args.map<ParameterInformation>(v => { return {label: v}})
+      });
     });
 
-    let prevSpace = line.lastIndexOf(' ', positionParams.position.character) + 1;
-    let keyword = line.substring(prevSpace, positionParams.position.character - 1);
-    let fn = functionCache.find(f => f.name.toLocaleLowerCase() === keyword.toLocaleLowerCase());
-    if (fn) {
-      fn.overloads.forEach(overload => {
-        signatures.push({
-          label: `${fn.name}( ${overload.args.join(', ')} )`,
-          documentation: overload.description,
-          parameters: overload.args.map<ParameterInformation>(v => { return {label: v}; })
-        })
-      });
-      activeSignature = 0;
-      activeParameter = 0;
-    }
+    activeSignature = 0;
+    activeParameter = Math.min(paramCount, overloads[0].args.length - 1);
   }
   return { signatures, activeSignature, activeParameter };
 })