+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to the "vscode-postgres" extension will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Syntax highlighting
+- Connection aware code completion (keywords, functions, tables, and fields)
+- In-line error detection powered by EXPLAIN (one error per query in editor)
+- *__Basic__* function signature support (connection aware)
 
 ## [0.0.6] - 2018-04-30
 ### Fixed